requirements.txt
python-telegram-bot==20.7
yfinance==0.2.28
pandas==2.2.2
feedparser==6.0.10
apscheduler==3.10.1
